requirements.txt
# Web scraping
playwright>=1.41.0
html5lib>=1.1

# Search engine
duckduckgo-search>=7.2.1

# LLM integration
openai>=1.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
anthropic>=0.8.0
google-generativeai>=0.3.0
openai[azure]>=1.0.0

# Testing
unittest2>=1.1.0
pytest>=7.0.0
pytest-asyncio>=0.23.5

# Google Generative AI
google-generativeai

# gRPC, for Google Generative AI preventing WARNING: All log messages before absl::InitializeLog() is called are written to STDERR
grpcio==1.60.1

# Data processing and visualization
yfinance>=0.2.36
pandas>=2.0.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Tabulate for pretty-printing tables
tabulate

# Utilities
requests>=2.25.0
uuid
beautifulsoup4>=4.9.0
tqdm>=4.65.0

# Added from the code block
numpy>=1.20.0